+++ src/app/pages/dentist/appointments-d/appointments-d.component.ts
@@ -233,7 +233,7 @@ export class AppointmentsDComponent implements OnInit {
 formatDateToLetter(date: any) {
   var startDate = new Date(date);
 
-  const formattedDate = new Intl.DateTimeFormat('es-ES', { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric'}).format(startDate);
+  const formattedDate = new Intl.DateTimeFormat('es-ES', { weekday: 'short', year: 'numeric', month: 'long', day: 'numeric'}).format(startDate);
 
   return formattedDate;
 }
@@ -241,7 +241,7 @@ formatDateToLetter(date: any) {
 formatDateToNumbers(date: any) {
   var startDate = new Date(date);
 
-  const formattedDate = new Intl.DateTimeFormat('es-ES', { hour: 'numeric', minute: 'numeric', weekday: 'long', year: 'numeric', month: 'long', day: 'numeric'}).format(startDate);
+  const formattedDate = new Intl.DateTimeFormat('es-ES', { hour: 'numeric', minute: 'numeric', weekday: 'short', year: 'numeric', month: 'long', day: 'numeric'}).format(startDate);
 
   return formattedDate;
 }
